Actualizando documento sobre generación de código.

+++ docs/codegen/c_binary_tree.c
@@ -1,5 +1,5 @@
 struct binary_tree {
-    int64_t value;
+    int value;
     struct binary_tree *left;
     struct binary_tree *right;
 }
